mexora_DatawereHouse/requirements.txt
pandas>=2.2.2
mysql-connector-python==8.4.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1